JENKINS-54397: Refactor classes and add documentation
